pyproject.toml
[tool.poetry]
name = "OpenTAKServer"
version = "0.0.0"
description = "A server for ATAK, WinTAK, and iTAK"
authors = ["OpenTAKServer <opentakserver@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"
repository = "https://github.com/brian7704/OpenTAKServer"
documentation = "https://docs.opentakserver.io"

[tool.poetry.dependencies]
adsbxcot = "6.0.4"
androguard = "4.1.2"
aiscot = "5.2.1"
beautifulsoup4 = "4.12.3"
colorlog = "6.8.2"
datetime = "5.5"
python-ffmpeg = "2.0.12"
flask = "3.0.3"
flask-apscheduler = "1.13.1"
flask-cors = "5.0.0"
flask-migrate = "4.0.7"
Flask-Security-Too = {version="5.5.2", extras=["common", "mfa"]}
flask-socketio = "5.3.7"
flask-sqlalchemy = "3.1.1"
eventlet = "0.37.0"
gevent = "24.2.1"
lastversion = "*"
lxml = "5.3.0"
meshtastic = "2.5.0"
pika = "1.3.2"
poetry-dynamic-versioning = {version = "1.4.1", extras = ["plugin"]}
protobuf = "5.28.2"
psutil = "6.0.0"
pyfiglet = "1.0.2"
pygc = "1.3.0"
PyJWT = "2.9.0"
pytak = "6.3.2"
pytest = "8.3.3"
pytest-cov = "4.1.0"
python = ">=3.10,<3.13"
python-socketio = {extras = ["client", "websocket_client", "asyncio_client"], version = "5.11.4"}
pyOpenSSL = "24.2.1"
pyotp = "2.9.0"
PyYAML = "6.0.2"
sqlalchemy = "2.0.35"
sqlalchemy-utils = "0.41.2"
tldextract = "5.1.2"
unishox2-py3 = "1.0.0"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
dirty = false
pattern = "((?P<epoch>\\d+)!)?(?P<base>\\d+(\\.\\d+)*)"

[tool.poetry-dynamic-versioning.files."opentakserver/__init__.py"]
persistent-substitution = true
initial-content = """
  # These version placeholders will be replaced later during substitution.
  __version__ = "0.0.0"
  __version_tuple__ = (0, 0, 0)
"""

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
opentakserver = "opentakserver.app:main"